requirements.txt
feedparser==6.0.11
pyaes==1.6.1
pyasn1==0.6.1
python-dotenv==1.1.1
rsa==4.9.1
sgmllib3k==1.0.0
Telethon==1.40.0
